requirements.txt
asgiref==3.8.1
Django==4.2
django-debug-toolbar==4.1.0
django-environ==0.11.2
pillow==11.1.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
django-debug-toolbar==4.1.0
djangorestframework==3.15.2
django-filter==24.3
Markdown==3.7